pyproject.toml
[tool.poetry]
name = "housing"
version = "0.1.0"
description = ""
authors = ["Alex Shadley <shadleyalex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
langchain = "^0.0.344"
openai = "^1.3.7"
tiktoken = "^0.5.1"
chromadb = "^0.4.18"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
python-multipart = "^0.0.6"
tika = "^2.6.0"
pypdf = "^3.17.1"
lxml = "^4.9.3"
beautifulsoup4 = "^4.12.2"
pinecone-client = "^2.2.4"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
